refactor: renamed imageUtils to urlUtils
